requirement.txt
flask
flask-script
flask-blueprint
flask-session
redis
flask-sqlAlchemy
pymysql
flask-login